mtd-utils: Correct casting for final status report in flashcp
Add correct casting for filestat.st_size in flashcp.c. While the
interim status updates had correct casting from commit 08b243, the
final update was not.

+++ misc-utils/flashcp.c
@@ -386,8 +386,8 @@ int main (int argc,char *argv[])
 	if (flags & FLAG_VERBOSE)
 		log_printf (LOG_NORMAL,
 				"\rVerifying data: %lluk/%lluk (100%%)\n",
-				KB (filestat.st_size),
-				KB (filestat.st_size));
+				KB ((unsigned long long)filestat.st_size),
+				KB ((unsigned long long)filestat.st_size));
 	DEBUG("Verified %d / %lluk bytes\n",written,(unsigned long long)filestat.st_size);
 
 	exit (EXIT_SUCCESS);
